+++ metaseq/__init__.py
@@ -19,7 +19,7 @@ __all__ = ["pdb"]
 # backwards compatibility to support `from metaseq.X import Y`
 from metaseq.distributed import utils as distributed_utils
 from metaseq.logging import meters, metrics  # noqa
-from .logging.progress_bar_dir import base_progress_bar
+from .logging.progress_bar import base_progress_bar
 
 sys.modules["metaseq.distributed_utils"] = distributed_utils
 sys.modules["metaseq.meters"] = meters
